+++ terraform/modules/alb/outputs.tf
@@ -27,3 +27,15 @@ output "target_group_arn" {
 
   value = aws_lb_target_group.this.arn
 }
+
+output "alb_arn_suffix" {
+  description = "ARN suffix of the Application Load Balancer."
+
+  value = aws_lb.this.arn_suffix
+}
+
+output "target_group_arn_suffix" {
+  description = "ARN suffix of the Target Group."
+
+  value = aws_lb_target_group.this.arn_suffix
+}