+++ terraform/modules/autoscaling/main.tf
@@ -3,18 +3,30 @@ resource "aws_autoscaling_group" "this" {
 
   vpc_zone_identifier = var.private_subnet_ids
 
-  desired_capacity = var.desired_capacity
-  min_size         = var.min_size
-  max_size         = var.max_size
+  desired_capacity      = var.desired_capacity
+  capacity_rebalance    = true
+  protect_from_scale_in = false
+  min_size              = var.min_size
+  max_size              = var.max_size
 
   health_check_type         = "ELB"
   health_check_grace_period = 300
-
-  target_group_arns = [var.target_group_arn]
+  default_instance_warmup   = 300
+  target_group_arns         = [var.target_group_arn]
 
   launch_template {
     id      = var.launch_template_id
     version = var.launch_template_latest_version
+  }
+
+  instance_refresh {
+    strategy = "Rolling"
+
+    preferences {
+      min_healthy_percentage = 50
+    }
+
+    triggers = ["launch_template"]
   }
 
   termination_policies = [