+++ terraform/modules/networking/main.tf
@@ -168,7 +168,7 @@ resource "aws_route_table" "public" {
 resource "aws_route_table" "private" {
   vpc_id = aws_vpc.this.id
 
-  route = {
+  route {
     cidr_block = "0.0.0.0/0"
 
     nat_gateway_id = aws_nat_gateway.this.id