+++ terraform/modules/networking/variables.tf
@@ -4,7 +4,7 @@ variable "project_name" {
 }
 
 variable "vpc_cidr" {
-  description = "CIDR block assigned to the VPS."
+  description = "CIDR block assigned to the VPC."
   type        = string
 }
 
@@ -24,7 +24,7 @@ variable "public_subnet_cidr_a" {
 }
 
 variable "public_subnet_cidr_b" {
-  description = "CIDR block for Public Subnet B"
+  description = "CIDR block for Public Subnet B."
   type        = string
 }
 
@@ -34,7 +34,7 @@ variable "private_subnet_cidr_a" {
 }
 
 variable "private_subnet_cidr_b" {
-  description = "CIDR block for Private Subnet B"
+  description = "CIDR block for Private Subnet B."
   type        = string
 }